Bump strimzi test-containers to 0.100.0
Add Kafka dev services with strimzi container with kraft
