+++ bom/application/pom.xml
@@ -139,6 +139,7 @@
         <kafka3.version>3.1.0</kafka3.version>
         <snappy.version>1.1.8.4</snappy.version>
         <zookeeper.version>3.5.7</zookeeper.version>
+        <strimzi-test-container.version>0.100.0</strimzi-test-container.version>
         <!-- Scala is used by Kafka so we need to choose a compatible version -->
         <scala.version>2.12.13</scala.version>
         <aws-lambda-java.version>1.2.1</aws-lambda-java.version>
@@ -3667,6 +3668,17 @@
                 <groupId>io.strimzi</groupId>
                 <artifactId>kafka-oauth-client</artifactId>
                 <version>${strimzi-oauth.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>io.strimzi</groupId>
+                <artifactId>strimzi-test-container</artifactId>
+                <version>${strimzi-test-container.version}</version>
+                <exclusions>
+                    <exclusion>
+                        <groupId>junit</groupId>
+                        <artifactId>junit</artifactId>
+                    </exclusion>
+                </exclusions>
             </dependency>
             <dependency>
                 <groupId>com.nimbusds</groupId>
