+++ bom/test/pom.xml
@@ -20,7 +20,7 @@
         <jaxb-api.version>2.3.1</jaxb-api.version>
 
         <rxjava1.version>1.3.8</rxjava1.version>
-        <strimzi-test-container.version>0.22.1</strimzi-test-container.version>
+        <strimzi-test-container.version>0.100.0</strimzi-test-container.version>
     </properties>
 
     <dependencyManagement>
@@ -46,6 +46,10 @@
                 <version>${strimzi-test-container.version}</version>
                 <scope>test</scope>
                 <exclusions>
+                    <exclusion>
+                        <groupId>org.slf4j</groupId>
+                        <artifactId>slf4j-simple</artifactId>
+                    </exclusion>
                     <exclusion>
                         <groupId>org.apache.logging.log4j</groupId>
                         <artifactId>log4j-core</artifactId>
