+++ integration-tests/kafka-oauth-keycloak/src/test/java/io/quarkus/it/kafka/KafkaKeycloakTestResource.java
@@ -1,18 +1,21 @@
 package io.quarkus.it.kafka;
 
+import static io.strimzi.test.container.StrimziKafkaContainer.KAFKA_PORT;
+
 import java.util.HashMap;
 import java.util.Map;
 
 import org.jboss.logging.Logger;
+import org.testcontainers.utility.MountableFile;
 
-import io.quarkus.it.kafka.containers.KafkaContainer;
 import io.quarkus.it.kafka.containers.KeycloakContainer;
 import io.quarkus.test.common.QuarkusTestResourceLifecycleManager;
+import io.strimzi.test.container.StrimziKafkaContainer;
 
 public class KafkaKeycloakTestResource implements QuarkusTestResourceLifecycleManager {
 
     private static final Logger log = Logger.getLogger(KafkaKeycloakTestResource.class);
-    private KafkaContainer kafka;
+    private StrimziKafkaContainer kafka;
     private KeycloakContainer keycloak;
 
     @Override
@@ -27,10 +30,16 @@ public class KafkaKeycloakTestResource implements QuarkusTestResourceLifecycleMa
         keycloak.createHostsFile();
 
         //Start kafka container
-        kafka = new KafkaContainer();
-        kafka.start();
-        log.info(kafka.getLogs());
-        properties.put("kafka.bootstrap.servers", kafka.getBootstrapServers());
+        this.kafka = new StrimziKafkaContainer("quay.io/strimzi/kafka:latest-kafka-3.0.0")
+                .withBrokerId(1)
+                .withKafkaConfigurationMap(Map.of("listener.security.protocol.map", "JWT:SASL_PLAINTEXT,BROKER1:PLAINTEXT"))
+                .withNetworkAliases("kafka")
+                .withServerProperties(MountableFile.forClasspathResource("kafkaServer.properties"))
+                .withBootstrapServers(
+                        c -> String.format("JWT://%s:%s", c.getContainerIpAddress(), c.getMappedPort(KAFKA_PORT)));
+        this.kafka.start();
+        log.info(this.kafka.getLogs());
+        properties.put("kafka.bootstrap.servers", this.kafka.getBootstrapServers());
 
         return properties;
     }
